move compact to its own file
